add 3d models for ALPS SKCL/SKCM switches
When comparing with the footprints, it seems like the offset of the
actualy pins of the switch and the footprint holes is not quite the
same. If the footprint is used with Mill-Max sockets, this might be a
problem. Most likely the 3d model is not quite right, but taking a quick
measurement of a real switch, it seems that possibly the holes might
also not be correct? Measuring the pin positions is not so easy, so
maybe that is the problem. Anyhow, the 3d models should be adjusted in
the future.